+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Added
+- Added a feed setting that allows feeds to specify if they'd like empty values to used to overwrite existing data. ([#1228](https://github.com/craftcms/feed-me/pull/1228), [#797](https://github.com/craftcms/feed-me/issues/797), [#723](https://github.com/craftcms/feed-me/issues/723), [#854](https://github.com/craftcms/feed-me/issues/854))
+
 ### Fixed
 - Fixed a bug where [direct feed](https://docs.craftcms.com/feed-me/v4/feature-tour/feed-overview.html#settings-pane) links in the control panel were malformed if [cpTrigger](https://craftcms.com/docs/4.x/config/general.html#cptrigger) was set to null. ([#1190](https://github.com/craftcms/feed-me/pull/1190))